third_party/blink/web_tests/fast/dom/HTMLIsIndexElement/prototype-chain-expected.txt

HTMLIsIndexElement should inherit directly from HTMLElement and be an HTMLUnknownElement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isIndex.__proto__ is HTMLUnknownElement.prototype
PASS isIndex.__proto__.__proto__ is HTMLElement.prototype
PASS isIndex.prompt is undefined.
PASS isIndex.form is undefined.
PASS isIndex.defaultValue is undefined.
PASS isIndex.disabled is undefined.
PASS isIndex.multiple is undefined.
PASS isIndex.alt is undefined.
PASS isIndex.accept is undefined.
PASS successfullyParsed is true

TEST COMPLETE
